+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # 0.1.1
 - add: some changes to m1
 - fix: wrong display of sprites of face anomalies
+
+# 0.1.2
+- add: changes and improvements to m1
+- add: custom materials for infected
+- add: custom sounds for infected
+- add: custom music
+- bugfixes